PC_host/requirements.txt
Flask==3.1.1
flask-socketio==5.4.1
paho-mqtt==2.1.0
python-socketio==5.13.2
eventlet==0.37.0
requests==2.32.4
blinker==1.9.0
python-dotenv==1.0.1